Create parse_json_file_test.zig
I needed to do this because the file_name's are locally relative.
But when tests are run "globally" the "full path" is needed. For example:

  var file_name = "modules/3d-test-resource/suzanne.babylon"

Needs to be changed to:

  var file_name = "src/modules/zig-json/modules/3d-test-resource/suzanne.babylon"

If I run the tests from "~/prgs/graphics/zig-3d-soft-engine/" with:

  $ make test

By splitting the tests out they are not run when tested globally because
they are not explicitly imported.